Rewrote how the Window is instantiated This now uses tobias-loew's [windows_thunks](https://github.com/tobias-loew/windows_thunks) to setup the Message Handler  - Encapsulates thunk creation and trampoline logic in a small library; fewer manual casts and safer lifetime handling  - Library provides tested helper to create per-window thunks and manage lifetime  - Move low-level details to submodule; easier to update and reuse  - Add submodule to build, include headers, link the thunk object/library  - Low: replace the setup thunk code with the library calls and adjust create/cleanup
This strategy maintains x86 and x64 compatibility, with debug and release builds:
 - Performance: per‑window thunk removes GWLP indirection and is minimal per‑message overhead — good for high‑frequency message pumps.
 - Complexity: you must manage executable memory (RW → RX), DEP, and safe deletion (free only after DestroyWindow returns). Be careful with lifetime and reentrancy; keep isDestroying_ guards in handlers.
 - Portability: thunk bytes differ for x86 vs x64; use a library that handles both architectures.

With this change, I cleaned up the WndProcHandlers for component classes, isolated the ImGui WndProcHandler evaluation so that it is handled in a clear way. I ended up providing a static pointer to the main window, and I may try to remove that in the future. It only exists right now to provide a way to process thread window messages with a nullptr hWnd.

To see this working, you can patch `imgui_demp.cpp` to execute the custom App Message to close the application using the Dear ImGui Demo → Menu → Quit menu item.

imgui_demo.cpp.patch:
```
diff --git "a/imgui_demo.cpp" "b/imgui_demo.cpp"
index 05b39a39f..c407e031e 100644
--- "a/imgui_demo.cpp"
+++ "b/imgui_demo.cpp"
@@ -114,6 +114,8 @@ Index of this file:
 #include <stdio.h>          // vsnprintf, sscanf, printf
 #include <stdlib.h>         // NULL, malloc, free, atoi
 #include <stdint.h>         // intptr_t
+
+#include "src/AtumWindows.hpp"
 #if !defined(_MSC_VER) || _MSC_VER >= 1800
 #include <inttypes.h>       // PRId64/PRIu64, not avail in some MinGW headers.
 #endif
@@ -8391,7 +8393,7 @@ static void ShowExampleMenuFile()
     }
     if (ImGui::MenuItem("Checked", NULL, true)) {}
     ImGui::Separator();
-    if (ImGui::MenuItem("Quit", "Alt+F4")) {}
+    if (ImGui::MenuItem("Quit", "Alt+F4")) { ::PostMessage(nullptr, WM_APP_CLOSE, WPARAM(0), LPARAM(0)); }
 }

 //-----------------------------------------------------------------------------

```

Improved the README.md documentation to make it easier to understand the project as it currently exists.